+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Lazy load DNS resolver to improve performance [PR #251](https://github.com/3scale/apicast/pull/251)
 
 ## [3.0.0-alpha2] - 2017-02-06
 ### Added